+++ scripts/publish-pages.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-git subtree push --prefix dist-ghpages origin gh-pages