+++ src/utils/isClient.spec.js
@@ -1,6 +1,0 @@
-import isClient from './isClient';
-
-describe('isClient utility', () => {
-  it('should be a boolean', () => expect(isClient).to.be.a('boolean'));
-  it('should return true during tests', () => expect(isClient).to.be.true);
-});