+++ test/isAPISupported.spec.js
@@ -1,0 +1,13 @@
+import isAPISupported from '../src/utils/isAPISupported';
+
+describe('isAPISupported utility', () => {
+  it('should be a function', () => {
+    expect(isAPISupported).to.be.a('function');
+  });
+
+  it('should return true if an API is supported', () => {
+    const result = isAPISupported('addEventListener');
+
+    expect(result).to.be.true;
+  });
+});