+++ test/isClient.spec.js
@@ -1,0 +1,6 @@
+import isClient from '../src/utils/isClient';
+
+describe('isClient utility', () => {
+  it('should be a boolean', () => expect(isClient).to.be.a('boolean'));
+  it('should return true during tests', () => expect(isClient).to.be.true);
+});